Docker & heroku
Got docker to set up. Took some time as there were tons of weird issues with uv and getting that all set up. I ended up having full Python image instead of the -slim, since it was having weird issues with dependencies.

That got it working and the docker container was set up. When deploying to heroku I keep getting an error when trying to commit to heroku.

Will spend more time tommorrow.